proveedor de servicios mejorado

+++ src/Core/Entities/Token.cs
@@ -5,7 +5,7 @@ namespace Cinte.Core.Entities
     public class Token
     {
         public string Token_Auth { get; set; }
-        public int Expire_In { get; set; }
+        public double Expire_In { get; set; }
         public string NickName { get; set; }
 
     }
